+++ util.h
@@ -8,6 +8,6 @@
 #define DOWNTO(l, n, m) (((0xffffffffU >> (31 - (n))) & l) >> (m))
 #define BIT(l, n) DOWNTO(l, n, n)
 float ui2f(uint32_t ui);
-float f2ui(float f);
+uint32_t f2ui(float f);
 
 #endif
